chore(app): fix some dev warnings
