+++ src/api/uic-inventory/src/components/Hooks/index.js
@@ -1,6 +1,8 @@
+export * from './useEditableInputs';
 export * from './useFetchHook';
 export * from './useGraphicManagerHook';
 export * from './useIntervalHook';
+export * from './useMaxLength';
 export * from './useOpenClosedHook';
 export * from './useSitePolygon';
 export * from './useViewPointZoomingHook';
