+++ src/api/uic-inventory/src/components/Hooks/useMaxLength.js
@@ -1,0 +1,6 @@
+export const useMaxLength = ({ value, limit }) => {
+  return {
+    limit,
+    remaining: limit - (value?.length || 0),
+  };
+};
